+++ src/policy/handcrafted/spell_profile.rs
@@ -364,6 +364,7 @@ impl HandcraftedPolicy {
             | EffectDef::ReduceGenericCostBy(_)
             | EffectDef::PlayersCantPlay(_)
             | EffectDef::LandwalkCanBeBlocked(_)
+            | EffectDef::CannotAttackUnless(_)
             | EffectDef::MultiplyEventAmount(_)
             | EffectDef::Replacement(_)
             | EffectDef::MoveToZone { .. }